[css-counter-styles] Implement complex predefined counter styles https://bugs.webkit.org/show_bug.cgi?id=256154 rdar://103021467
Reviewed by Darin Adler.

Implement the following as complex counter style systems:

simp-chinese-informal
simp-chinese-formal
trad-chinese-informal
trad-chinese-formal
ethiopic-numeric

As a result, authors can now create counter styles that extend the above.

Specification: https://drafts.csswg.org/css-counter-styles/#complex-predefined-counters

To implement this, we create UA sheet only values for the system descriptors.
The algorithms are just moved over from RenderListMarker, and re-used in both places as static methods on CSSCounterStyle.

* LayoutTests/TestExpectations:
* Source/WebCore/css/CSSCounterStyle.cpp: Move over code from RenderListMarker
(WebCore::counterForSystemCJK):
(WebCore::CSSCounterStyle::counterForSystemSimplifiedChineseInformal):
(WebCore::CSSCounterStyle::counterForSystemSimplifiedChineseFormal):
(WebCore::CSSCounterStyle::counterForSystemTraditionalChineseInformal):
(WebCore::CSSCounterStyle::counterForSystemTraditionalChineseFormal):
(WebCore::CSSCounterStyle::counterForSystemEthiopicNumeric):
(WebCore::CSSCounterStyle::initialRepresentation const):
(WebCore::CSSCounterStyle::isInRange const):
* Source/WebCore/css/CSSCounterStyle.h:
* Source/WebCore/css/CSSCounterStyleDescriptors.cpp: Pipe the CSS value for system descriptor
(WebCore::CSSCounterStyleDescriptors::areSymbolsValidForSystem):
(WebCore::CSSCounterStyleDescriptors::systemCSSText const):
* Source/WebCore/css/CSSCounterStyleDescriptors.h:
* Source/WebCore/css/CSSCounterStyleRegistry.cpp: Enable the new codepath for complex system
(WebCore::isCounterStyleUnsupportedByUserAgent):
* Source/WebCore/css/CSSCounterStyleRule.cpp: Pipe the CSS value for system descriptor
(WebCore::toCounterStyleSystemEnum):
* Source/WebCore/css/CSSProperties.json: Pass in the context, so the parser knows whether it's in UA sheet mode
* Source/WebCore/css/CSSValueKeywords.in: Add internal CSS values
* Source/WebCore/css/counterStyles.css: Define the new complex counter styles
(@counter-style simp-chinese-informal):
(@counter-style simp-chinese-formal):
(@counter-style trad-chinese-informal):
(@counter-style trad-chinese-formal):
(@counter-style cjk-ideographic):
(@counter-style ethiopic-numeric):
* Source/WebCore/css/parser/CSSPropertyParserHelpers.cpp: Parse the new internal CSS values
(WebCore::CSSPropertyParserHelpers::consumeCounterStyleSystem): Remove outdated FIXME that is handled by CSSCounterStyleDescriptors::areSymbolsValidForSystem
* Source/WebCore/css/parser/CSSPropertyParserHelpers.h:
* Source/WebCore/rendering/RenderListMarker.cpp: Delete code that was moved to CSSCounterStyle, and re-use the methods there
(WebCore::listMarkerText):
(WebCore::toCJKIdeographic): Deleted.
(WebCore::toEthiopicNumeric): Deleted.

Canonical link: https://commits.webkit.org/263544@main
