+++ Source/WebCore/rendering/RenderListMarker.cpp
@@ -55,7 +55,6 @@ WTF_MAKE_ISO_ALLOCATED_IMPL(RenderListMarker);
 constexpr int cMarkerPadding = 7;
 
 enum class SequenceType : bool { Numeric, Alphabetic };
-enum class Formality : bool { Informal, Formal };
 
 template<SequenceType type, typename CharacterType> static String toAlphabeticOrNumeric(int number, const CharacterType* sequence, unsigned sequenceSize)
 {
@@ -135,91 +134,6 @@ template<typename CharacterType, size_t size> static String toSymbolic(int numbe
     return toSymbolic(number, alphabet, size);
 }
 
-// This table format was derived from an old draft of the CSS specification: 3 group markers, 3 digit markers, 10 digits, negative sign.
-static String toCJKIdeographic(int number, const UChar table[17], Formality formality)
-{
-    enum AbstractCJKCharacter {
-        NoChar,
-        SecondGroupMarker, ThirdGroupMarker, FourthGroupMarker,
-        SecondDigitMarker, ThirdDigitMarker, FourthDigitMarker,
-        Digit0, Digit1, Digit2, Digit3, Digit4,
-        Digit5, Digit6, Digit7, Digit8, Digit9,
-        NegativeSign
-    };
-
-    if (!number)
-        return { &table[Digit0 - 1] , 1 };
-
-    ASSERT(number != std::numeric_limits<int>::min());
-    bool needsNegativeSign = number < 0;
-    if (needsNegativeSign)
-        number = -number;
-
-    constexpr unsigned groupLength = 8; // 4 digits, 3 digit markers, and a group marker
-    constexpr unsigned bufferLength = 4 * groupLength;
-    AbstractCJKCharacter buffer[bufferLength] = { NoChar };
-
-    for (int i = 0; i < 4; ++i) {
-        int groupValue = number % 10000;
-        number /= 10000;
-
-        // Process least-significant group first, but put it in the buffer last.
-        auto group = &buffer[(3 - i) * groupLength];
-
-        if (groupValue && i)
-            group[7] = static_cast<AbstractCJKCharacter>(SecondGroupMarker - 1 + i);
-
-        // Put in the four digits and digit markers for any non-zero digits.
-        group[6] = static_cast<AbstractCJKCharacter>(Digit0 + (groupValue % 10));
-        if (number != 0 || groupValue > 9) {
-            int digitValue = ((groupValue / 10) % 10);
-            group[4] = static_cast<AbstractCJKCharacter>(Digit0 + digitValue);
-            if (digitValue)
-                group[5] = SecondDigitMarker;
-        }
-        if (number != 0 || groupValue > 99) {
-            int digitValue = ((groupValue / 100) % 10);
-            group[2] = static_cast<AbstractCJKCharacter>(Digit0 + digitValue);
-            if (digitValue)
-                group[3] = ThirdDigitMarker;
-        }
-        if (number != 0 || groupValue > 999) {
-            int digitValue = groupValue / 1000;
-            group[0] = static_cast<AbstractCJKCharacter>(Digit0 + digitValue);
-            if (digitValue)
-                group[1] = FourthDigitMarker;
-        }
-
-        if (formality == Formality::Informal && groupValue < 20) {
-            // Remove the tens digit, but leave the marker.
-            ASSERT(group[4] == NoChar || group[4] == Digit0 || group[4] == Digit1);
-            group[4] = NoChar;
-        }
-
-        if (!number)
-            break;
-    }
-
-    // Convert into characters, omitting consecutive runs of digit0 and trailing digit0.
-    unsigned length = 0;
-    UChar characters[1 + bufferLength];
-    auto last = NoChar;
-    if (needsNegativeSign)
-        characters[length++] = table[NegativeSign - 1];
-    for (unsigned i = 0; i < bufferLength; ++i) {
-        auto character = buffer[i];
-        if (character != NoChar) {
-            if (character != Digit0 || last != Digit0)
-                characters[length++] = table[character - 1];
-            last = character;
-        }
-    }
-    if (last == Digit0)
-        --length;
-
-    return { characters, length };
-}
-
 struct AdditiveSymbol {
     int value;
     std::initializer_list<UChar> characters;
@@ -256,48 +170,6 @@ static String toPredefinedAdditiveSystem(int value, const AdditiveSystem& system
     }
     ASSERT(!value);
     return result.toString();
-}
-
-static String toEthiopicNumeric(int value)
-{
-    ASSERT(value >= 1);
-
-    if (value == 1) {
-        UChar ethiopicDigitOne = 0x1369;
-        return { &ethiopicDigitOne, 1 };
-    }
-
-    // Split the number into groups of two digits, starting with the least significant decimal digit.
-    uint8_t groups[5];
-    for (auto& group : groups) {
-        group = value % 100;
-        value /= 100;
-    }
-
-    UChar buffer[std::size(groups) * 3];
-    unsigned length = 0;
-    bool isMostSignificantGroup = true;
-    for (int i = std::size(groups) - 1; i >= 0; --i) {
-        auto value = groups[i];
-        bool isOddIndex = i & 1;
-        // If the group has the value zero, or if the group is the most significant one and has the value 1,
-        // or if the group has an odd index (as given in the previous step) and has the value 1,
-        // then remove the digits (but leave the group, so it still has a separator appended below).
-        if (!(value == 1 && (isMostSignificantGroup || isOddIndex))) {
-            if (auto tens = value / 10)
-                buffer[length++] = 0x1371 + tens;
-            if (auto ones = value % 10)
-                buffer[length++] = 0x1368 + ones;
-        }
-        if (value && isOddIndex)
-            buffer[length++] = 0x137B;
-        if ((value || !isMostSignificantGroup) && !isOddIndex && i)
-            buffer[length++] = 0x137C;
-        if (value)
-            isMostSignificantGroup = false;
-    }
-
-    return { buffer, length };
 }
 
 static ListStyleType::Type effectiveListMarkerType(ListStyleType::Type type, int value)
@@ -953,49 +825,15 @@ String listMarkerText(ListStyleType::Type type, int value, CSSCounterStyle* coun
         };
         return toAlphabetic(value, upperNorwegianAlphabet);
     }
-
-    case ListStyleType::Type::SimplifiedChineseInformal: {
-        static constexpr UChar simplifiedChineseInformalTable[17] = {
-            0x842C, 0x5104, 0x5146, // These three group markers are probably wrong; OK because we don't use this on big enough numbers.
-            0x5341, 0x767E, 0x5343,
-            0x96F6, 0x4E00, 0x4E8C, 0x4E09, 0x56DB,
-            0x4E94, 0x516D, 0x4E03, 0x516B, 0x4E5D,
-            0x8D1F
-        };
-        return toCJKIdeographic(value, simplifiedChineseInformalTable, Formality::Informal);
-    }
-    case ListStyleType::Type::SimplifiedChineseFormal: {
-        static constexpr UChar simplifiedChineseFormalTable[17] = {
-            0x842C, 0x5104, 0x5146, // These three group markers are probably wrong; OK because we don't use this on big enough numbers.
-            0x62FE, 0x4F70, 0x4EDF,
-            0x96F6, 0x58F9, 0x8D30, 0x53C1, 0x8086,
-            0x4F0D, 0x9646, 0x67D2, 0x634C, 0x7396,
-            0x8D1F
-        };
-        return toCJKIdeographic(value, simplifiedChineseFormalTable, Formality::Formal);
-    }
+    case ListStyleType::Type::SimplifiedChineseInformal:
+        return CSSCounterStyle::counterForSystemSimplifiedChineseInformal(value);
+    case ListStyleType::Type::SimplifiedChineseFormal:
+        return CSSCounterStyle::counterForSystemSimplifiedChineseFormal(value);
     case ListStyleType::Type::CJKIdeographic:
-    case ListStyleType::Type::TraditionalChineseInformal: {
-        static constexpr UChar traditionalChineseInformalTable[17] = {
-            0x842C, 0x5104, 0x5146,
-            0x5341, 0x767E, 0x5343,
-            0x96F6, 0x4E00, 0x4E8C, 0x4E09, 0x56DB,
-            0x4E94, 0x516D, 0x4E03, 0x516B, 0x4E5D,
-            0x8CA0
-        };
-        return toCJKIdeographic(value, traditionalChineseInformalTable, Formality::Informal);
-    }
-    case ListStyleType::Type::TraditionalChineseFormal: {
-        static constexpr UChar traditionalChineseFormalTable[17] = {
-            0x842C, 0x5104, 0x5146, // These three group markers are probably wrong; OK because we don't use this on big enough numbers.
-            0x62FE, 0x4F70, 0x4EDF,
-            0x96F6, 0x58F9, 0x8CB3, 0x53C3, 0x8086,
-            0x4F0D, 0x9678, 0x67D2, 0x634C, 0x7396,
-            0x8CA0
-        };
-        return toCJKIdeographic(value, traditionalChineseFormalTable, Formality::Formal);
-    }
-
+    case ListStyleType::Type::TraditionalChineseInformal:
+        return CSSCounterStyle::counterForSystemTraditionalChineseInformal(value);
+    case ListStyleType::Type::TraditionalChineseFormal:
+        return CSSCounterStyle::counterForSystemTraditionalChineseFormal(value);
     case ListStyleType::Type::LowerRoman: {
         static CONSTEXPR_WITH_MSVC_INITIALIZER_LIST_WORKAROUND AdditiveSystem lowerRomanSystem {
             {
@@ -1460,7 +1298,7 @@ String listMarkerText(ListStyleType::Type type, int value, CSSCounterStyle* coun
     }
 
     case ListStyleType::Type::EthiopicNumeric:
-        return toEthiopicNumeric(value);
+        return CSSCounterStyle::counterForSystemEthiopicNumeric(value);
 
     case ListStyleType::Type::String:
         ASSERT_NOT_REACHED();
